refactor(animate): move to script setup
