+++ src/packages/__VUE/animate/types.ts
